Refactor imports in BuyTokenPage and TokenDetails components; add fromWei utility function for value conversion
